+++ spacemacs-themes/ewal-spacemacs-themes.el
@@ -1,7 +1,8 @@
 ;;; ewal-spacemacs-themes.el --- Ride the rainbow spaceship -*- lexical-binding: t; -*-
 
-;; Version: 0.3.1
-;; Package-Requires: ((emacs "25") (ewal "0.3.0") (spacemacs-theme "0.2.0"))
+;; Version: 0.3.2
+;; Package-Requires: ((emacs "25.1") (ewal "0.3.0") (spacemacs-theme "0.2.0"))
+;; URL: https://github.com/fkr-0/ewal
 
 ;;; Commentary:
 ;;